metadataのみ取得できるようにする
